moved viewModel interfaces from infrastructure
